FARRAGO:
Re-org SqlOperatorTests into a more TestCase class, with a non-static method for each test. It's abstract, to allow a derived class to provide a tester, or to add more tests.

I've removed the SqlOperator.test(SqlTester) method which I was previously insisting on because it provided linkage between operators and their tests. Now the linkage goes the other way: a test must call SqlTester.isFor(SqlOperator) when it starts.

Still to be done: Write a SqlTester which ensures that every operator has a minimum number of tests. Maybe it could generate a report. Anyway, assuming that this would be written, I took out a lot of TODOs.

[git-p4: depot-paths = "//open/dt/dev/": change = 3792]